refactor: move extension entrypoints
