+++ .changeset/restructure-extension-entrypoints.md
@@ -1,0 +1,5 @@
+---
+"@aliou/pi-synthetic": patch
+---
+
+Move extension entry points to the top-level `extensions/` directory and centralize non-provider Synthetic API calls in a shared client.
